Move test data to the right place.
This file seems to have been forgotten when the test it belongs to got
moved -- probably because the test was poorly written and still passed
when the file couldn't be opened.  The file is supposed to live one
directory up from the qtextbrowser test so that it can be used to test
opening url's with relative paths.

Change-Id: Id80965b1c73b37de75691ae3a18f909f9c6d799a
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>