Enable use of QArrayDataPointer as a return type
This is done through a (POD) wrapper around a raw QTypedArrayData
pointer. The wrapper forces conscious decisions on the use of that
pointer, while implying it is internal API.

The pointed-to header is assumed to referenced, as if already owned by
the receiving QArrayDataPointer.

This wrapper is placed in the generic qarraydata.h header as its
definition doesn't depend on QArrayDataPointer by itself.

Change-Id: I9b2b6bb7dd9ab073dc7d882ccf999b32d793ed3f
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>