Removed include to qdatastream.h from qmetatype.h
Added forward reference to QDataStream instead

In the new connection syntax, I want to include qmetatype.h from qobject.h,
in order to handle metatype for queued connection.
So qdatastream cannot be included here (as it inherit qobject) else it would
lead to circular inclusion.

QDataStream is only used in templated code that is dependent on the
template parametter so it is fine to have it forward declared.

Change-Id: I08ea479f481262a80451a277ee29533c42cb9f89
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>