Allow overriding the output dir when running syncqt
This feature is only available to advanced qmake clients who have their
own defaults_pre.prf, where QMAKE_SYNCQT_OUTDIR is set before loading
Qt's own default_pre.

This is useful to be able to put the sync.profile file in a different
level in the source tree than where you want the output in the build
tree.

An alternate/complimenting solution would be to allow the output path
to be overriden from within the sync.profile file.

Change-Id: I3cdef82dbe492877507e269e18571456beeb2b7d
Reviewed-by: Marius Storm-Olsen <marius.storm-olsen@nokia.com>