Build all tests on Mac OS X, but do not run them yet.
Previously, autotests and benchmarks were disabled on Mac OS X. This is
wrong, and needs to be corrected. The 'make check' target is disabled
for the autotests on Mac OS X for now, but all tests will now be built.
The check target does not actually run the benchmarks, so this can be
safely re-enabled without breaking the CI system.

There will be commits to follow which re-enable the 'make check' target
on Mac OS X for each module.

Change-Id: I1347da46e3688b48e8cc5d933ee5188ea3723639
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>