Ensure qmodule.pri is installed.
From the commit message of 0b2ce8520020a0db9b5b5feb5e72f64518d3eea5 and
the code in qt_module.prf, it is apparently intended that qmodule.pri is
installed, but this was never implemented correctly.  Make sure we
install it.

This has gone unnoticed because it happens to be installed by accident
unless $$QT_SOURCE_TREE != $$QT_BUILD_TREE and
$$QT_BUILD_TREE != $$[QT_INSTALL_PREFIX] (i.e. a shadow build which is
not using -developer-build).

Change-Id: Iee861a7bb592ca43a61ad91f1ef6a7a5bd21aff8
Reviewed-by: Jyri Tahtela <jyri.tahtela@nokia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>