syncqt: fixed QT_MODULE_BASE when build path != install path
QT_MODULE_BASE is exclusively used to refer to the top-level source
directory of a module.  However, in non-developer-build mode (used
only when the build path is not equal to the install path), syncqt
would incorrectly set the value of QT_MODULE_BASE to the build
directory.  This caused all $$QT.<module>.sources variables to point
to the wrong place.

Make it always point to the source directory.

Change-Id: I401cf000f9e3a4ab97d68db40330bb98aec25a9a
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>
Reviewed-by: Marius Storm-Olsen <marius.storm-olsen@nokia.com>